+++ src/main/java/com/komandda/web/controller/rest/SeriesController.java
@@ -4,6 +4,7 @@ import com.komandda.entity.Event;
 import com.komandda.entity.User;
 import com.komandda.service.LocationService;
 import com.komandda.service.SeriesService;
+import com.komandda.web.controller.rest.permission.PermissionChecker;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -29,21 +30,26 @@ public class SeriesController {
     @Autowired
     private SimpMessagingTemplate webSocket;
 
-    @PreAuthorize("hasAuthority('event_edit')")
+    @Autowired
+    private PermissionChecker permissionChecker;
+
+    @PreAuthorize("hasAnyAuthority('event_edit', 'self_event_edit')")
     @RequestMapping(method = RequestMethod.POST)
     public void add(@RequestBody Event event, @AuthenticationPrincipal User user) {
         webSocket.convertAndSend("/event/create", service.insert(event, user));
     }
 
-    @PreAuthorize("hasAuthority('event_edit')")
+    @PreAuthorize("hasAnyAuthority('event_edit', 'self_event_edit')")
     @RequestMapping(method = RequestMethod.PUT)
     public void update(@RequestBody Event event, @AuthenticationPrincipal User user) {
+        permissionChecker.checkSelfEditPermission(event, user);
         webSocket.convertAndSend("/event/update", service.save(event, user));
     }
 
-    @PreAuthorize("hasAuthority('event_edit')")
+    @PreAuthorize("hasAnyAuthority('event_edit', 'self_event_edit')")
     @RequestMapping(method = RequestMethod.DELETE)
     public void delete(@RequestBody Event event, @AuthenticationPrincipal User user) {
+        permissionChecker.checkSelfEditPermission(event, user);
         webSocket.convertAndSend("/event/delete", service.delete(event, user));
     }
 }
